Accepts to download the model in a different folder #11

+++ qrdet/qrdet.py
@@ -25,7 +25,8 @@ _MODEL_FILE_NAME = 'qrdet-{size}.pt'
 
 
 class QRDetector:
-    def __init__(self, model_size: str = 's', conf_th: float = 0.5, nms_iou: float = 0.3):
+    def __init__(self, model_size: str = 's', conf_th: float = 0.5, nms_iou: float = 0.3,
+                 weights_folder: str = _WEIGHTS_FOLDER):
         """
         Initialize the QRDetector.
         It loads the weights of the YOLOv8 model and prepares it for inference.
@@ -39,6 +40,9 @@ class QRDetector:
         assert model_size in ('n', 's', 'm', 'l'), f'Invalid model size: {model_size}. ' \
                                                    f'Valid values are: \'n\', \'s\', \'m\' or \'l\'.'
         self._model_size = model_size
+        self.weights_folder = weights_folder
+        self.__current_release_txt_file = os.path.join(weights_folder, 'current_release.txt')
+
         path = self.__download_weights_or_return_path(model_size=model_size)
         assert os.path.exists(path), f'Could not find model weights at {path}.'
 
@@ -104,19 +108,19 @@ class QRDetector:
         :param desc: str. The description of the download. Default: 'Downloading weights...'.
         """
         self.downloading_model = True
-        path = os.path.join(_WEIGHTS_FOLDER, _MODEL_FILE_NAME.format(size=model_size))
+        path = os.path.join(self.weights_folder, _MODEL_FILE_NAME.format(size=model_size))
         if os.path.isfile(path):
-            if os.path.isfile(_CURRENT_RELEASE_TXT_FILE):
+            if os.path.isfile(self.__current_release_txt_file):
                 # Compare the current release with the actual release URL
-                with open(_CURRENT_RELEASE_TXT_FILE, 'r') as file:
+                with open(self.__current_release_txt_file, 'r') as file:
                     current_release = file.read()
                 # If the current release is the same as the URL, the weights are already downloaded.
                 if current_release == _WEIGHTS_URL_FOLDER:
                     self.downloading_model = False
                     return path
         # Create the directory to save the weights.
-        elif not os.path.exists(_WEIGHTS_FOLDER):
-            os.makedirs(_WEIGHTS_FOLDER)
+        elif not os.path.exists(self.weights_folder):
+            os.makedirs(self.weights_folder)
 
         url = f"{_WEIGHTS_URL_FOLDER}/{_MODEL_FILE_NAME.format(size=model_size)}"
 
@@ -134,7 +138,7 @@ class QRDetector:
                     progress_bar.update(len(data))
                     file.write(data)
         # Save the current release URL
-        with open(_CURRENT_RELEASE_TXT_FILE, 'w') as file:
+        with open(self.__current_release_txt_file, 'w') as file:
             file.write(_WEIGHTS_URL_FOLDER)
         # Check the weights were downloaded correctly.
         if total_size_in_bytes != 0 and progress_bar.n != total_size_in_bytes:
@@ -146,7 +150,8 @@ class QRDetector:
         return path
 
     def __del__(self):
-        path = os.path.join(_WEIGHTS_FOLDER, _MODEL_FILE_NAME.format(size=self._model_size))
-        # If the weights didn't finish downloading, delete them.
-        if hasattr(self, 'downloading_model') and self.downloading_model and os.path.isfile(path):
-            os.remove(path)
+        if hasattr(self, 'weights_folder'):
+            path = os.path.join(self.weights_folder, _MODEL_FILE_NAME.format(size=self._model_size))
+            # If the weights didn't finish downloading, delete them.
+            if hasattr(self, 'downloading_model') and self.downloading_model and os.path.isfile(path):
+                os.remove(path)
